Move inheritance tests into the proper file
