+++ tests/test_callbacks.py
@@ -1,32 +1,6 @@
 import pytest
 
 
-def test_inheritance(msg):
-    from pybind11_tests import Pet, Dog, Rabbit, dog_bark, pet_name_species
-
-    roger = Rabbit('Rabbit')
-    assert roger.name() + " is a " + roger.species() == "Rabbit is a parrot"
-    assert pet_name_species(roger) == "Rabbit is a parrot"
-
-    polly = Pet('Polly', 'parrot')
-    assert polly.name() + " is a " + polly.species() == "Polly is a parrot"
-    assert pet_name_species(polly) == "Polly is a parrot"
-
-    molly = Dog('Molly')
-    assert molly.name() + " is a " + molly.species() == "Molly is a dog"
-    assert pet_name_species(molly) == "Molly is a dog"
-
-    assert dog_bark(molly) == "Woof!"
-
-    with pytest.raises(TypeError) as excinfo:
-        dog_bark(polly)
-    assert msg(excinfo.value) == """
-        Incompatible function arguments. The following argument types are supported:
-            1. (arg0: m.Dog) -> str
-            Invoked with: <m.Pet object at 0>
-    """
-
-
 def test_callbacks():
     from functools import partial
     from pybind11_tests import (test_callback1, test_callback2, test_callback3,
